+++ tests/persistence/read.rs
@@ -3,19 +3,20 @@ use std::sync::Arc;
 
 use worktable::prelude::*;
 
-use crate::persistence::{get_empty_test_wt, get_test_wt, TestWorkTable, TEST_INNER_SIZE, TEST_PAGE_SIZE};
+// TODO: Fix naming.
+use crate::persistence::{get_empty_test_wt, get_test_wt, TestPersistWorkTable, TESTPERSIST_INNER_SIZE, TESTPERSIST_PAGE_SIZE};
 
 #[test]
 fn test_info_parse() {
-    let mut file = File::open("tests/data/expected/test.wt").unwrap();
-    let info = parse_page::<SpaceInfoData, { TEST_PAGE_SIZE as u32 }>(&mut file, 0).unwrap();
+    let mut file = File::open("tests/data/expected/test_persist.wt").unwrap();
+    let info = parse_page::<SpaceInfoData, { TESTPERSIST_INNER_SIZE as u32 }>(&mut file, 0).unwrap();
 
     assert_eq!(info.header.space_id, 0.into());
     assert_eq!(info.header.page_id, 0.into());
     assert_eq!(info.header.previous_id, 0.into());
     assert_eq!(info.header.next_id, 1.into());
     assert_eq!(info.header.page_type, PageType::SpaceInfo);
-    assert_eq!(info.header.data_length, 108);
+    assert_eq!(info.header.data_length, 120);
 
     assert_eq!(info.inner.id, 0.into());
     assert_eq!(info.inner.page_count, 2);
@@ -35,8 +36,8 @@ fn test_info_parse() {
 
 #[test]
 fn test_index_parse() {
-    let mut file = File::open("tests/data/expected/test.wt").unwrap();
-    let index = parse_page::<IndexData<u128>, { TEST_PAGE_SIZE as u32 }>(&mut file, 1).unwrap();
+    let mut file = File::open("tests/data/expected/test_persist.wt").unwrap();
+    let index = parse_page::<IndexData<u128>, { TESTPERSIST_PAGE_SIZE as u32 }>(&mut file, 1).unwrap();
 
     assert_eq!(index.header.space_id, 0.into());
     assert_eq!(index.header.page_id, 1.into());
@@ -68,8 +69,8 @@ fn test_index_parse() {
 
 #[test]
 fn test_data_parse() {
-    let mut file = File::open("tests/data/expected/test.wt").unwrap();
-    let data = parse_data_page::<{ TEST_PAGE_SIZE }, { TEST_INNER_SIZE }>(&mut file, 3).unwrap();
+    let mut file = File::open("tests/data/expected/test_persist.wt").unwrap();
+    let data = parse_data_page::<{ TESTPERSIST_PAGE_SIZE }, { TESTPERSIST_INNER_SIZE }>(&mut file, 3).unwrap();
 
     assert_eq!(data.header.space_id, 0.into());
     assert_eq!(data.header.page_id, 3.into());
@@ -85,7 +86,7 @@ fn test_space_parse() {
         config_path: "tests/data".to_string(),
         database_files_dir: "tests/data/expected".to_string(),
     });
-    let table = TestWorkTable::load_from_file(manager).unwrap();
+    let table = TestPersistWorkTable::load_from_file(manager).unwrap();
     let expected = get_test_wt();
 
     assert_eq!(
@@ -100,7 +101,7 @@ fn test_space_parse_no_file() {
         config_path: "tests/data".to_string(),
         database_files_dir: "tests/data/non-existent".to_string(),
     });
-    let table = TestWorkTable::load_from_file(manager).unwrap();
+    let table = TestPersistWorkTable::load_from_file(manager).unwrap();
     let expected = get_empty_test_wt();
     assert_eq!(
         table.select_all().execute().unwrap(),